pyproject.toml
[project]
name = "music-downloader"
version = "0.1.0"
description = ""
authors = [
    {name = "Temerold", email = "tobbe@temerold.se"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp (>=2025.9.26,<2026.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
]

[tool.poetry]
packages = [{include = "music_downloader", from = "src"}]

[tool.poetry.scripts]
download = "music_downloader.cli:download"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
